Add OBEX support
Credit to the fantastic work done in Blueman
by Christopher Schramm and Sander Sweers.

This is a python Bluez obex agent ported from Blueman.
It is launched via XDG autostart and enabled via gsettings (org.blueberry obex-enabled).

Its process name is blueberry-obex-agent. It uses notifications to approve/reject
file transfers.

At the time of Mint 18.1, the situation in regards to OBEX support wasn't ideal:

- Gnome-user-share 3.14's OBEX support isn't functional and disappears in version 3.18 (same code, moved to gnome-bluetooth).
- The gnome-bluetooth 3.18 widget used in Blueberry is supposed to support OBEX but it isn't functional.
- The gnome-bluetooth 3.18 widget is tied to its parent but OBEX should work whether blueberry itself is running or not.

This commit makes OBEX work ootb.

Enabling/disabling that support is done via gsettings followed by a logout/login.

Whether to expose this setting in the UI and/or to ship with OBEX disabled by default was
not decided yet.