Overwrite the label in GnomeBluetooth SettingsWidget
A few things are wrong with the widget:

- It shows the name of the adapter instead of its alias
- It shows the state of gnome-bt obex (which isn't functional) instead of the state of blueberry's obex support
- In case of misconfiguration or race condition it doesn't get set at all,
which leads to showing its default value in Glade (the notorious "Bastien's computer" label and a link
making Downloads point to gnome.org).

Blueberry now tries to introspect the content of the widget,
removes the shadows on the scrolledwindow,
and sets the label appropriately.