blueberry: rework rfkill error to be switch-specific
instead of using the status stack page which prevents use of the
bluetooth widget, we now show an error icon next to the switch
and disable it while not changing any other state. If bluetooth
is blocked then we end up in a bluetooth disabled + switch disabled
state, otherwise we end up in a normal view with the switch disabled.