Handle error from rfkill
Fixes #88

In the case where `/dev/rfkill` does not exist `rfkill` will return
false which triggers a backtrace in blueberry, which may then
trigger a distro's bug reporting tool. This is common on virtual
machines.

Instead handle the error. `res` gets set to an empty string, which
the rest of the funtion interprets and having no adapter. This allows
blueberry to launch with a window saying

"No Bluetooth adapters found"